refactor(pack): refactor xdsl diagnostic
