+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/diagnostic/details/details.html
