+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/diagnostic/details/details.routing.js
@@ -1,0 +1,15 @@
+import controller from './details.controller';
+import template from './details.html';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state(
+    'telecom.packs.pack.xdsl.line.access-diagnostic-details',
+    {
+      url: '/diagnostic-details',
+      template,
+      controller,
+      controllerAs: '$ctrl',
+      layout: 'modal',
+    },
+  );
+};
