+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/diagnostic/diagnostic.controller.js
@@ -1,0 +1,6 @@
+export default /* @ngInject */ class XdslDiagnosticCtrl {
+  constructor($stateParams) {
+    this.lineNumber = $stateParams.number;
+    this.serviceName = $stateParams.serviceName;
+  }
+}
