--- packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/diagnostic/pack-xdsl-access-diagnostic.controller.js
@@ -1,9 +1,0 @@
-angular.module('managerApp').controller(
-  'XdslDiagnosticCtrl',
-  class XdslDiagnosticCtrl {
-    constructor($stateParams) {
-      this.lineNumber = $stateParams.number;
-      this.serviceName = $stateParams.serviceName;
-    }
-  },
-);
